+++ syntho_hive/__init__.py
@@ -1,3 +1,3 @@
 from .interface.config import Metadata, PrivacyConfig
 
-__version__ = "1.2.2"
+__version__ = "1.2.3"